requirements.txt
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.26.3
pathlib==1.0.1
boto3==1.34.34
botocore==1.34.34
s3fs==2024.2.0 